// src/Api/models/HomeCategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HomeCategory
 */
export interface HomeCategory {
    /**
     * 
     * @type {number}
     * @memberof HomeCategory
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    image?: string;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    categoryId?: string;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    section?: HomeCategorySectionEnum;
}


/**
 * @export
 */
export const HomeCategorySectionEnum = {
    ElectricCategories: 'ELECTRIC_CATEGORIES',
    Grid: 'GRID',
    ShopByCategories: 'SHOP_BY_CATEGORIES',
    Deals: 'DEALS'
} as const;
export type HomeCategorySectionEnum = typeof HomeCategorySectionEnum[keyof typeof HomeCategorySectionEnum];


/**
 * Check if a given object implements the HomeCategory interface.
 */
export function instanceOfHomeCategory(value: object): value is HomeCategory {
    return true;
}

export function HomeCategoryFromJSON(json: any): HomeCategory {
    return HomeCategoryFromJSONTyped(json, false);
}

export function HomeCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): HomeCategory {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'image': json['image'] == null ? undefined : json['image'],
        'categoryId': json['categoryId'] == null ? undefined : json['categoryId'],
        'section': json['section'] == null ? undefined : json['section'],
    };
}

  export function HomeCategoryToJSON(json: any): HomeCategory {
      return HomeCategoryToJSONTyped(json, false);
  }

  export function HomeCategoryToJSONTyped(value?: HomeCategory | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'image': value['image'],
        'categoryId': value['categoryId'],
        'section': value['section'],
    };
}