// src/Api/models/home-category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface HomeCategory
 */
export interface HomeCategory {
    /**
     * 
     * @type {number}
     * @memberof HomeCategory
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    'image'?: string;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    'categoryId'?: string;
    /**
     * 
     * @type {string}
     * @memberof HomeCategory
     */
    'section'?: HomeCategorySectionEnum;
}

export const HomeCategorySectionEnum = {
    ElectricCategories: 'ELECTRIC_CATEGORIES',
    Grid: 'GRID',
    ShopByCategories: 'SHOP_BY_CATEGORIES',
    Deals: 'DEALS'
} as const;

export type HomeCategorySectionEnum = typeof HomeCategorySectionEnum[keyof typeof HomeCategorySectionEnum];